Adds markdown support
