+++ next.config.mjs
@@ -1,0 +1,29 @@
+import remarkFrontmatter from 'remark-frontmatter';
+import mdx from '@next/mdx';
+/** @type {import('next').NextConfig} */
+
+const withMDX = mdx({
+  extension: /\.mdx?$/,
+  options: {
+    remarkPlugins: [remarkFrontmatter],
+  },
+  providerImportSource: '@mdx-js/react',
+});
+
+const nextConfig = {
+  reactStrictMode: true,
+  compiler: {
+    styledComponents: true,
+  },
+  webpack(config) {
+    config.module.rules.push({
+      test: /\.svg$/i,
+      issuer: /\.[jt]sx?$/,
+      use: ['@svgr/webpack'],
+    });
+    return config;
+  },
+  pageExtensions: ['ts', 'tsx', 'md', 'mdx'],
+};
+
+export default withMDX(nextConfig);
