+++ package.json
@@ -33,6 +33,7 @@
     "next": "13.1.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "remark-frontmatter": "^4.0.1",
     "styled-components": "^5.3.6",
     "typescript": "4.9.4"
   },
