+++ scripts/generate-blog-list.mjs
@@ -3,10 +3,18 @@ import matter from 'gray-matter';
 
 const fileNames = fs.readdirSync('./src/pages/blog');
 
+const fileNamesToIgnore = ['index.tsx'];
+
 const blogListMetaDataArray = fileNames.reduce((blogListMetaData, fileName) => {
-  const fileContents = fs.readFileSync(`./src/pages/blog/${fileName}`, 'utf-8');
-  const { data } = matter(fileContents);
-  return [...blogListMetaData, data];
+  if (!fileNamesToIgnore.includes(fileName)) {
+    const fileContents = fs.readFileSync(
+      `./src/pages/blog/${fileName}`,
+      'utf-8',
+    );
+    const { data, content } = matter(fileContents);
+    return [...blogListMetaData, { data, content }];
+  }
+  return [...blogListMetaData];
 }, []);
 
 fs.writeFile(`bloglist.json`, JSON.stringify(blogListMetaDataArray), () => {
